% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expenditure}
\alias{expenditure}
\title{Expenditure}
\format{
\subsection{\code{expenditure}}{

\describe{
\item{lga_name}{Local government area name}
\item{financial_year}{Financial year}
\item{measure}{Measure description}
\item{value}{Value}
...
}
}
}
\usage{
expenditure
}
\description{
A tibble with 4 columns and 3541 rows.
}
\keyword{datasets}
