% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facts}
\alias{facts}
\title{Facts}
\format{
\subsection{\code{facts}}{

\describe{
\item{lga_name}{Local government area name}
\item{financial_year}{Financial year}
\item{measure}{Measure description}
\item{value}{Value}
...
}
}
}
\usage{
facts
}
\description{
A tibble with 4 columns and 790 rows.
}
\keyword{datasets}
