% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kpis}
\alias{kpis}
\title{KPIs}
\format{
\subsection{\code{kpis}}{

\describe{
\item{lga_name}{Local government area name}
\item{financial_year}{Financial year}
\item{measure}{Measure description}
\item{value}{Value}
\item{source}{Source of data}
...
}
}
}
\usage{
kpis
}
\description{
A tibble with 5 columns and 8366 rows.
}
\keyword{datasets}
