% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lgv_rents}
\alias{lgv_rents}
\title{Rents}
\format{
\subsection{\code{population}}{
}
}
\source{
https://www.dffh.vic.gov.au/publications/rental-report
\describe{
rental_type
\item{rental_type}{The type of rental property, 1 bedroom, 2 bedroom, All properties}
\item{lga_name}{Local government area name}
\item{rental_date}{End of quarter that data is for, eg 30 June 2021}
\item{value}{A numeric value.}
\item{measure}{The type of data either the median rental value or count of rentals}
...
}
}
\usage{
lgv_rents
}
\description{
A tibble with 5 columns and 111706 rows.
}
\keyword{datasets}
