% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revenue}
\alias{revenue}
\title{Revenue}
\format{
\subsection{\code{revenue}}{

\describe{
\item{lga_name}{Local government area name}
\item{financial_year}{Financial year}
\item{measure}{Measure description}
\item{value}{Value}
...
}
}
}
\usage{
revenue
}
\description{
A tibble with 5 columns and 3931 rows.
}
\keyword{datasets}
